Title: 씨 오 더블유, Time: 0ms, Memory: 0MB, Status: Passed - Codetree

+++ 260419/씨 오 더블유/c-o-w.java
@@ -12,7 +12,7 @@ public class Main {
     		arr[i] = line.charAt(i - 1);
     	}
     	
-    	int[][] prefix = new int[3][N + 1];
+    	long[][] prefix = new long[3][N + 1];
     	HashMap<Integer, Character> map = new HashMap<>();
     	map.put(0, 'C');
     	map.put(1, 'O');
